refactor: use kebab-case for new editor modules
