+++ apps/desktop/src/routes/editor/Timeline/image-track.tsx
@@ -1,4 +1,4 @@
-import { OverlayTrack, type OverlayTrackProps } from "./StyleTrack";
+import { OverlayTrack, type OverlayTrackProps } from "./style-track";
 
 export function ImageTrack(props: OverlayTrackProps) {
 	return <OverlayTrack {...props} type="image" />;
