+++ apps/desktop/src/routes/editor/image-overlay.tsx
