+++ apps/desktop/src/routes/editor/image-segment-config.tsx
