+++ apps/desktop/src/routes/editor/style-segment-config.tsx
